+++ src/go/doc/testdata/examples/empty.golden
@@ -1,0 +1,6 @@
+-- .Play --
+package main
+
+func main() {}
+func main()
+
