+++ src/go/doc/testdata/examples/whole_function_external.go
@@ -1,0 +1,12 @@
+// Copyright 2021 The Go Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+package foo_test
+
+func foo(int)
+
+func Example() {
+	foo(42)
+	// Output:
+}
