+++ src/go/doc/testdata/examples/whole_function_external.golden
@@ -1,0 +1,8 @@
+-- .Play --
+package main
+
+func foo(int)
+
+func main() {
+	foo(42)
+}
